linear optimizer directory reorganized

+++ lib/src/classifier/linear/softmax_regressor/softmax_regressor_impl.dart
@@ -5,7 +5,6 @@ import 'package:ml_algo/src/helpers/get_probabilities.dart';
 import 'package:ml_algo/src/linear_optimizer/linear_optimizer.dart';
 import 'package:ml_algo/src/link_function/link_function.dart';
 import 'package:ml_algo/src/predictor/assessable_predictor_mixin.dart';
-import 'package:ml_algo/src/utils/parameter_default_values.dart';
 import 'package:ml_linalg/dtype.dart';
 import 'package:ml_linalg/matrix.dart';
 
@@ -20,7 +19,7 @@ class SoftmaxRegressorImpl with LinearClassifierMixin,
         bool fitIntercept = false,
         double interceptScale = 1.0,
         Matrix initialWeights,
-        this.dtype = ParameterDefaultValues.dtype,
+        this.dtype = DType.float32,
       }) :
         fitIntercept = fitIntercept,
         interceptScale = interceptScale,
